requirements-dev.txt
a2wsgi==1.10.8
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
alembic==1.15.2
amqp==5.3.1
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
asttokens==3.0.0
asyncpg==0.30.0
attrs==25.3.0
azure-ai-documentintelligence==1.0.2
azure-ai-inference==1.0.0b9
azure-core==1.33.0
azure-core-tracing-opentelemetry==1.0.0b12
azure-cosmos==4.9.0
azure-identity==1.21.0
azure-keyvault-secrets==4.9.0
backoff==2.2.1
bcrypt==4.3.0
beautifulsoup4==4.13.4
billiard==4.2.1
blinker==1.9.0
branca==0.8.1
build==1.2.2.post1
cachetools==5.5.2
celery==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.1
chroma-hnswlib==0.7.6
chromadb==1.0.9
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.2
contourpy==1.3.2
cryptography==44.0.2
cycler==0.12.1
dataclasses-json==0.6.7
debugpy==1.8.14
decorator==5.2.1
Deprecated==1.2.18
dill==0.4.0
distro==1.9.0
dnspython==2.7.0
docling==2.31.1
docling-core==2.27.0
docling-ibm-models==3.4.1
docling-parse==4.0.1
duckduckgo_search==8.0.1
durationpy==0.9
easyocr==1.7.2
et_xmlfile==2.0.0
executing==2.2.0
fastapi==0.115.9
filelock==3.18.0
filetype==1.2.0
Flask==3.1.0
Flask-Login==0.6.3
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
flatbuffers==25.2.10
flower==2.0.1
folium==0.19.5
fonttools==4.58.0
frozenlist==1.5.0
fsspec==2025.3.0
future==1.0.0
geographiclib==2.0
geopy==2.4.1
google-auth==2.39.0
googleapis-common-protos==1.70.0
greenlet==3.1.1
grpcio==1.71.0
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.23.5
humanfriendly==10.0
humanize==4.12.2
hyperframe==6.1.0
idna==3.10
imageio==2.37.0
importlib_metadata==8.6.1
importlib_resources==6.5.2
ipykernel==6.29.5
ipython==9.2.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.7
isodate==0.7.2
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.9.0
joblib==1.4.2
jsonlines==3.1.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonref==1.1.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_widgets==3.0.15
kiwisolver==1.4.8
kombu==5.5.3
kubernetes==32.0.1
langchain==0.3.27
langchain-azure-ai==0.1.2
langchain-chroma==0.2.4
langchain-community==0.3.27
langchain-core==0.3.74
langchain-docling==1.1.0
langchain-experimental==0.3.4
langchain-huggingface==0.1.2
langchain-openai==0.3.30
langchain-postgres==0.0.14
langchain-text-splitters==0.3.9
langgraph==0.4.1
langgraph-checkpoint==2.0.25
langgraph-prebuilt==0.1.8
langgraph-sdk==0.1.66
langgraph-supervisor==0.0.19
langsmith==0.4.15
lark==1.2.2
latex2mathml==3.77.0
lazy_loader==0.4
lxml==5.3.2
Mako==1.3.10
markdown-it-py==3.0.0
marko==2.1.3
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.3
matplotlib-inline==0.1.7
mdurl==0.1.2
mmh3==5.1.0
monotonic==1.6
mpire==2.10.2
mpmath==1.3.0
msal==1.32.0
msal-extensions==1.3.1
multidict==6.2.0
multiprocess==0.70.18
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.4.2
ninja==1.11.1.4
numpy==1.26.2
# nvidia-cublas-cu12==12.4.5.8  # Disabled for macOS/Linux CPU environments
# nvidia-cuda-cupti-cu12==12.4.127  # Disabled for macOS/Linux CPU environments
# nvidia-cuda-nvrtc-cu12==12.4.127  # Disabled for macOS/Linux CPU environments
# nvidia-cuda-runtime-cu12==12.4.127  # Disabled for macOS/Linux CPU environments
# nvidia-cudnn-cu12==9.1.0.70  # Disabled for macOS/Linux CPU environments
# nvidia-cufft-cu12==11.2.1.3  # Disabled for macOS/Linux CPU environments
# nvidia-curand-cu12==10.3.5.147  # Disabled for macOS/Linux CPU environments
# nvidia-cusolver-cu12==11.6.1.9  # Disabled for macOS/Linux CPU environments
# nvidia-cusparse-cu12==12.3.1.170  # Disabled for macOS/Linux CPU environments
# nvidia-cusparselt-cu12==0.6.2  # Disabled for macOS/Linux CPU environments
# nvidia-nccl-cu12==2.21.5  # Disabled for macOS/Linux CPU environments
# nvidia-nvjitlink-cu12==12.4.127  # Disabled for macOS/Linux CPU environments
# nvidia-nvtx-cu12==12.4.127  # Disabled for macOS/Linux CPU environments
o365==2.1.2
oauthlib==3.2.2
onnxruntime==1.21.1
openai==1.100.2
opencv-python-headless==4.11.0.86
openpyxl==3.1.5
opentelemetry-api==1.32.1
opentelemetry-exporter-otlp-proto-common==1.32.1
opentelemetry-exporter-otlp-proto-grpc==1.32.1
opentelemetry-instrumentation==0.53b1
opentelemetry-instrumentation-asgi==0.53b1
opentelemetry-instrumentation-fastapi==0.53b1
opentelemetry-proto==1.32.1
opentelemetry-sdk==1.32.1
opentelemetry-semantic-conventions==0.53b1
opentelemetry-util-http==0.53b1
orjson==3.10.18
ormsgpack==1.9.1
outcome==1.3.0.post0
overrides==7.7.0
packaging==24.2
pandas==2.2.3
parso==0.8.4
pdfminer.six==20250416
pexpect==4.9.0
pgvector==0.3.6
pillow==11.1.0
platformdirs==4.3.7
pluggy==1.5.0
posthog==4.0.0
primp==0.15.0
priority==2.0.0
prometheus_client==0.21.1
prompt_toolkit==3.0.51
propcache==0.3.0
protobuf==5.29.4
psutil==7.0.0
psycopg[binary]==3.2.6
psycopg-pool==3.2.6
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybind11==2.13.6
pyclipper==1.3.0.post6
pycparser==2.22
pydantic==2.11.4
pydantic-settings==2.8.1
pydantic_core==2.33.2
Pygments==2.19.1
PyJWT==2.10.1
pylatexenc==2.10
pymongo==4.12.1
PyMuPDF==1.25.5
pyparsing==3.2.3
PyPDF2==3.0.1
pypdfium2==4.30.1
PyPika==0.48.9
pyproject_hooks==1.2.0
PySocks==1.7.1
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-pptx==1.0.2
pytz==2025.2
PyYAML==6.0.2
pyzmq==26.4.0
rapidocr-onnxruntime==1.4.4
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.0.0
rpds-py==0.24.0
rsa==4.9.1
rtree==1.4.0
safetensors==0.5.3
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.2
selenium==4.32.0
semchunk==2.2.2
sentence-transformers
setuptools==80.7.1
shapely==2.1.0
shellingham==1.5.4
simsimd==6.2.1
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
SQLAlchemy==2.0.39
stack-data==0.6.3
starlette==0.45.3
sympy==1.13.1
tabulate==0.9.0
tenacity==9.1.2
threadpoolctl==3.6.0
tifffile==2025.3.30
tiktoken==0.9.0
# tokenizers==0.21.1  # Disabled to resolve conflict between transformers and chromadb, using compatible version below
tokenizers==0.19.1
# torch==2.6.0  # Disabled for macOS/Linux CPU environments, use compatible version below
torch==2.2.2
# torchvision==0.21.0  # Disabled for macOS/Linux CPU environments, use compatible version below
torchvision==0.17.2
tornado==6.4.2
tqdm==4.66.4
traitlets==5.14.3
# transformers==4.50.0  # Disabled for macOS due to conflict with docling-ibm-models, using compatible version below
transformers==4.42.4
trio==0.30.0
trio-websocket==0.12.2
# triton==3.2.0  # Disabled for macOS/Linux CPU environments, not available for this Python version
typer==0.15.2
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.4.0
vine==5.1.0
watchfiles==1.0.5
wcwidth==0.2.13
webdriver-manager==4.0.2
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
widgetsnbextension==4.0.14
wrapt==1.17.2
wsproto==1.2.0
WTForms==3.2.1
XlsxWriter==3.2.3
xxhash==3.5.0
xyzservices==2025.4.0
yarl==1.18.3
zipp==3.21.0
zstandard==0.23.0
poml
langdetect
gunicorn>=20.1.0